requirements.txt
django==3.1
djangorestframework==3.12
psycopg2-binary
flake8==3.6.0
